refactored App to be in main src directory instead of components
